Hello,
Now a days Flipkart is Providing the Login Feature with EmailID with OTP
or Phone Number with OTP By using Selenium OTP verification is not
possible.

so I have Integrated the 3rd Party API (Mailosaur)
After integrated this I Have Enter the Email and extract the OTP aslo
But The FlipKart App is checks whether the user is Human or not.

After Enter the OTP and click on continue it asks click and HOld.

so this is not possible to Keep Hold by Selenium.

In This Project I have Tried to Login with Email ID and OTP.